Fixed bug with deleting the solution.
Solutions are displayed as numbers, and the DEL button uses the
substr function to delete from the screen so I had to add toString.

There's still the issue that it deletes the whole solution instead of
the last character. This happens because the solutions are inserted as
whole numbers rather than individual numbers side by side. Might fix it,
might not. Not really a big deal.